[feat] 도서 대여 기능 구현

+++ src/main/java/com/programmers/library/io/Input.java
@@ -3,7 +3,7 @@ package com.programmers.library.io;
 public interface Input {
 	int inputMode();
 	int inputMenu();
-	int inputBookId();
+	long inputBookId();
 	String inputBookTitle();
 	String inputBookAuthor();
 	long inputBookPages();
